src/plataform/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
cssbeautifier==1.14.9
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.1
djlint==1.34.0
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.9
json5==0.9.14
oauthlib==3.2.2
pathspec==0.11.2
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
djangorestframework==3.14.0
